backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.6.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.0
# Preserving previously required deps for auth to work
python-dotenv
email-validator
bcrypt
passlib
